MapViewModelをインタフェースと実装に分離
