+++ app/src/main/java/com/example/foos/ui/view/screen/map/MapScreen.kt
@@ -47,7 +47,7 @@ import kotlinx.coroutines.launch
 @OptIn(ExperimentalPermissionsApi::class, ExperimentalMaterialApi::class)
 @SuppressLint("MissingPermission")
 @Composable
-fun MapScreen(viewModel: MapViewModel, navController: NavController) {
+fun MapScreen(viewModel: MapViewModelImpl, navController: NavController) {
 
     val uiState = viewModel.uiState.value
